Combine selection and input handling logic for beat snap grids across all rulesets
